Consolidate stats worker implementation
