+++ tests/test_stats_focus_async.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from PySide6.QtCore import Qt
 
 from Tools.Stats.PySide6.stats_ui_pyside6 import StatsWindow
-from Tools.Stats.PySide6.stats_worker import StatsWorker
+from Tools.Stats.PySide6.stats_workers import StatsWorker
 
 
 def test_stats_focus_async(qtbot, monkeypatch):
