+++ tests/test_stats_window_smoke_phase0.py
@@ -6,7 +6,7 @@ from PySide6.QtCore import Qt
 
 from Tools.Stats.PySide6 import stats_ui_pyside6 as stats_mod
 from Tools.Stats.PySide6.stats_ui_pyside6 import StatsWindow
-from Tools.Stats.PySide6.stats_worker import StatsWorker
+from Tools.Stats.PySide6.stats_workers import StatsWorker
 
 
 @pytest.fixture
